Initial commit of reading genres and genre playlists from echonest
the proof is in this here pudding: we display a stack of genre cards and
let the user swipe through them. once they "swipe right" on one, we
lookup a playlist for that genre and allow them to swipe through those
songs.

that's all this does right now. basic proof of concept that we can pull
the info we need from echonest.