Explicitly shuffle genre collection on initial load
for some odd reason, on heroku, we're always returning the same shuffled
collection meaning the genres are always in the same order. barring
divine intervention, something is probably off with our code. it could
possibly be the minifier/optimizer as i can't reproduce locally, but
this should fix it up at least